+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/dto/GlucoseDto.java
@@ -2,7 +2,6 @@ package gr.codehub.sacchon.app.dto;
 
 import gr.codehub.sacchon.app.model.Glucose;
 import gr.codehub.sacchon.app.model.Measurement;
-import gr.codehub.sacchon.app.model.Patient;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,7 +17,7 @@ import java.util.stream.Collectors;
 @Getter
 public class GlucoseDto implements Measurement<List<GlucoseRecordDto>> {
 
-    private int id;
+    private Long id;
     private final String UNITS = "mg/dL";
     private LocalDate date;
     private List<GlucoseRecordDto> measurement;
