+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/repository/CarbsRepository.java
@@ -11,17 +11,17 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface CarbsRepository extends JpaRepository<Carbs, Integer> {
+public interface CarbsRepository extends JpaRepository<Carbs, Long> {
 
     @Query(value = "SELECT * FROM " + SacchonApplication.SCHEMA + ".CARBS WHERE CARBS.PATIENT_ID = :patientId",
             nativeQuery = true)
-    List<Carbs> findCarbsByPatientId(@Param("patientId") int patientId);
+    List<Carbs> findCarbsByPatientId(@Param("patientId") long patientId);
 
     @Query(value = "SELECT AVG(MEASUREMENT) FROM " + SacchonApplication.SCHEMA + ".CARBS " +
             "WHERE (CARBS.PATIENT_ID = :patientId AND CARBS.DATE >= :startingDate AND CARBS.DATE <= :endingDate)",
             nativeQuery = true)
     Integer findAverageCarbsByPatientIdOnSpecificDates(
-            @Param("patientId") int patientId,
+            @Param("patientId") long patientId,
             @Param("startingDate") LocalDate startingDate,
             @Param("endingDate") LocalDate endingDate
     );
