+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/repository/ConsultationRepository.java
@@ -9,11 +9,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ConsultationRepository extends JpaRepository<Consultation, Integer> {
+public interface ConsultationRepository extends JpaRepository<Consultation, Long> {
 
     @Query(value = "SELECT * FROM " + SacchonApplication.SCHEMA + ".CONSULTATION WHERE CONSULTATION.PATIENT_ID = :patientId", nativeQuery = true)
-    List<Consultation> findConsultationByPatientId(@Param("patientId") int patientId);
+    List<Consultation> findConsultationByPatientId(@Param("patientId") long patientId);
 
     @Query(value = "SELECT * FROM " + SacchonApplication.SCHEMA + ".CONSULTATION WHERE CONSULTATION.PATIENT_ID = :patientId", nativeQuery = true)
-    List<Consultation> findConsultationInfoByPatientId(@Param("patientId") int patientId);
+    List<Consultation> findConsultationInfoByPatientId(@Param("patientId") long patientId);
 }
