+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/service/CarbsService.java
@@ -11,10 +11,10 @@ public interface CarbsService {
 
     CarbsDto createCarbsIntake(CarbsDto carbsDto);
     List<CarbsDto> readCarbs();
-    CarbsDto readCarbsById(int id) throws CarbsException;
-    List<CarbsFromPersonDto> readCarbsByPatientId(int id);
-    Integer readAverageCarbsIntakeByPatientIdOnSpecificDates(int id, LocalDate startingDate, LocalDate endingDate);
-    boolean updateCarbs(CarbsDto carbsDto, int id) throws CarbsException;
-    boolean deleteCarbsById(int id) throws CarbsException;
+    CarbsDto readCarbsById(long id) throws CarbsException;
+    List<CarbsFromPersonDto> readCarbsByPatientId(long id);
+    Integer readAverageCarbsIntakeByPatientIdOnSpecificDates(long id, LocalDate startingDate, LocalDate endingDate);
+    boolean updateCarbs(CarbsDto carbsDto, long id) throws CarbsException;
+    boolean deleteCarbsById(long id) throws CarbsException;
 
 }
