+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/service/ChiefDoctorService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface ChiefDoctorService {
     ChiefDoctorDto createChiefDoctor(ChiefDoctorDto chiefDoctor);
     List<ChiefDoctorDto> readChiefDoctor();
-    ChiefDoctorDto readChiefDoctor(int id) throws ChiefDoctorException;
-    boolean updateChiefDoctor(ChiefDoctorDto chiefDoctor, int id);
-    boolean deleteChiefDoctor(int id);
+    ChiefDoctorDto readChiefDoctor(long id) throws ChiefDoctorException;
+    boolean updateChiefDoctor(ChiefDoctorDto chiefDoctor, long id);
+    boolean deleteChiefDoctor(long id);
 }
